Create javadoc.java
javadoc.java is an advanced program comprised of a high-tech sorting alg. that allows for a deck to be sorted if it is not already sorted. It is one of the few programs that can be applied in a real life scenerio and I would appreciate if you asked me BEFORE using my code.